Fixed double/int generation. More datas.

+++ road.hpp
@@ -20,7 +20,7 @@ class road {
   std::size_t size_out() const;
   bool empty_in() const;
   bool empty_out() const;
-  void newcar_rd(bool const, int, const int, const double);
+  void newcar_rd(bool const, double, const int, const double);
   void evolve_rd(bool const, rbout&, double const,double const, double, double, double, const double, const double);
   bool transfer_rd();
   void erase_rd();
